Doom scrolling is bad.

+++ src/classes/point.cpp
@@ -101,10 +101,10 @@ Point::Point(Cartesian2D cart, double r, double az, double alt, double m)
 // https://stackoverflow.com/questions/849211/shortest-distance-between-a-point-and-a-line-segment
 double Point::get_distance_to_line(Point a, Point b) const
 {
-    float r2 = pow(a.distance_to(b), 2);
+    double r2 = pow(a.distance_to(b), 2);
     if (!r2) return distance_to(a);
 
-    float t = fmax(0, fmin(1,  ((x - a.x) * (b.x - a.x) + (y - a.y) * (b.y - a.y) + (z - a.z) * (b.z - a.z)) / r2));
+    double t = fmax(0, fmin(1,  ((x - a.x) * (b.x - a.x) + (y - a.y) * (b.y - a.y) + (z - a.z) * (b.z - a.z)) / r2));
     Point p(a.x + t * (b.x-a.x), a.y + t * (b.y-a.y), a.z + t * (b.z-a.z));
 
     return distance_to(p);
